Extract trajectory analysis from HTML reporting
